# contracts/openapi/mediaflow.v1.yaml
openapi: 3.1.2
info:
  title: MediaFlow API
  version: v1
  description: Contract-first API for MediaFlow identity, inbox scanning, continuous discovery, deterministic identification, review administration, formal catalog reads, downloader management, and source automation.
servers:
  - url: /
paths:
...
  /api/v1/integrations/tmdb:
    get:
      operationId: getTmdbIntegration
      summary: Read TMDB integration configuration and health
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/TmdbIntegration' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: putTmdbIntegration
      summary: Replace TMDB integration configuration
      security: *write-security
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/Origin' }
        - { $ref: '#/components/parameters/SecFetchSite' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody: { $ref: '#/components/requestBodies/PutTmdbIntegrationRequest' }
      responses:
        '200': { $ref: '#/components/responses/TmdbIntegration' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteTmdbIntegration
      summary: Delete TMDB integration credentials
      security: *write-security
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/Origin' }
        - { $ref: '#/components/parameters/SecFetchSite' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '204': { description: TMDB integration deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/integrations/tmdb/connection-tests:
    post:
      operationId: testTmdbConnection
      summary: Test TMDB credentials without saving them
      security: *write-security
      parameters: *write-parameters
      requestBody: { $ref: '#/components/requestBodies/TmdbConnectionTestRequest' }
      responses:
        '200': { $ref: '#/components/responses/TmdbConnectionTestResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/v1/identification-enhancer:
    get:
      operationId: getIdentificationEnhancer
      summary: Read the redacted local identification enhancer configuration
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/IdentificationEnhancer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: putIdentificationEnhancer
      summary: Replace the local identification enhancer configuration
      security: *write-security
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/Origin' }
        - { $ref: '#/components/parameters/SecFetchSite' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody: { $ref: '#/components/requestBodies/IdentificationEnhancerInput' }
      responses:
        '200': { $ref: '#/components/responses/IdentificationEnhancer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/identification-enhancer/connection-tests:
    post:
      operationId: testIdentificationEnhancer
      summary: Test a candidate local Ollama endpoint without saving it
      security: *write-security
      parameters: *write-parameters
      requestBody: { $ref: '#/components/requestBodies/IdentificationEnhancerInput' }
      responses:
        '200': { $ref: '#/components/responses/IdentificationEnhancerConnectionTestResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: __Host-mediaflow_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
...
    CsrfToken: { name: X-CSRF-Token, in: header, required: true, schema: { type: string, minLength: 1 } }
    Origin: { name: Origin, in: header, required: false, description: Exact public origin proof used when Fetch Metadata is absent and checked whenever supplied., schema: { type: string, format: uri } }
    SecFetchSite: { name: Sec-Fetch-Site, in: header, required: false, description: Browser Fetch Metadata primary same-origin proof; cross-site and malformed values fail closed., schema: { type: string, enum: [same-origin, same-site, cross-site, none] } }
...
    IfMatch: { name: If-Match, in: header, required: true, description: Decimal configuration version expected by the caller., schema: { type: string, pattern: '^[0-9]{1,19}$' } }
...
  requestBodies:
...
    PutTmdbIntegrationRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PutTmdbIntegrationRequest' } } } }
    TmdbConnectionTestRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TmdbConnectionTestRequest' } } } }
...
    IdentificationEnhancerInput: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/IdentificationEnhancerInput' } } } }
...
  responses:
...
    TmdbIntegration: { description: Redacted TMDB configuration and health, content: { application/json: { schema: { $ref: '#/components/schemas/TmdbIntegration' } } } }
    TmdbConnectionTestResult: { description: Non-persistent TMDB connection test result, content: { application/json: { schema: { $ref: '#/components/schemas/TmdbConnectionTestResult' } } } }
...
    IdentificationEnhancer: { description: Redacted local enhancer configuration, content: { application/json: { schema: { $ref: '#/components/schemas/IdentificationEnhancer' } } } }
    IdentificationEnhancerConnectionTestResult: { description: Non-persistent local enhancer test result, content: { application/json: { schema: { $ref: '#/components/schemas/IdentificationEnhancerConnectionTestResult' } } } }
...
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    Forbidden: { description: CSRF or Origin rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
...
    Conflict: { description: Request conflicts with current state, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    ValidationFailed: { description: Request validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    RateLimited: { description: Request rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    ServiceUnavailable: { description: Service is not ready, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    InternalError: { description: Internal error, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
  schemas:
    ApiErrorBody:
      description: Allow-listed API failure envelope with a correlation request ID and optional safe structured details.
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code: { type: string, enum: [bootstrap.already_completed, bootstrap.invalid_secret, auth.invalid_credentials, auth.rate_limited, session.expired, csrf.invalid, origin.untrusted, root.not_found, root.unavailable, path.invalid, path.escape, path.symlink_forbidden, inbox.overlap, inbox.not_found, task.not_found, task.invalid_state, task.lease_lost, integration.not_configured, integration.unauthorized, integration.rate_limited, integration.unsupported-version, provider.unavailable, provider.timeout, provider.response-too-large, provider.invalid-response, download.correlation-ambiguous, download.remote-missing, automation.source-disabled, automation.signature-invalid, automation.replay, automation.action-invalid, automation.payload-invalid, automation.downstream-conflict, organization.root-read-only, organization.target-overlap, organization.target-unavailable, organization.path-outside-root, organization.target-exists, organization.source-changed, organization.hardlink-cross-device, organization.operation-unsupported, organization.journal-ambiguous, revision.changed, event.cursor_expired, request.conflict, validation.failed, internal.error] }
            message: { type: string }
            request_id: { type: string, format: uuid }
            details: { type: object, additionalProperties: { type: [string, number, boolean, 'null'] } }
...
    TmdbIntegration:
      description: Redacted TMDB configuration and health projection; credentials and ciphertext are never returned.
      type: object
      additionalProperties: false
      required: [kind, configured, locale, region, config_version, health, checked_at, failure_code]
      properties:
        kind: { const: tmdb }
        configured: { type: boolean }
        locale: { type: string, pattern: '^[a-z]{2}-[A-Z]{2}$', default: zh-CN }
        region: { type: [string, 'null'], pattern: '^[A-Z]{2}$' }
        config_version: { type: integer, minimum: 0, maximum: 9007199254740991 }
        health: { enum: [unconfigured, healthy, degraded, unavailable, unauthorized, rate-limited] }
        checked_at: { type: [string, 'null'], format: date-time }
        failure_code: { oneOf: [{ $ref: '#/components/schemas/IntegrationFailureCode' }, { type: 'null' }] }
    PutTmdbIntegrationRequest:
      description: TMDB API Read Access Token and non-sensitive locale preferences.
      type: object
      additionalProperties: false
      required: [api_read_access_token, locale, region]
      properties:
        api_read_access_token: { type: string, minLength: 16, maxLength: 4096, writeOnly: true }
        locale: { type: string, pattern: '^[a-z]{2}-[A-Z]{2}$', default: zh-CN }
        region: { type: [string, 'null'], pattern: '^[A-Z]{2}$' }
    TmdbConnectionTestRequest:
      description: Candidate TMDB credentials used only for a non-persistent connection test.
      type: object
      additionalProperties: false
      required: [api_read_access_token, locale, region]
      properties:
        api_read_access_token: { type: string, minLength: 16, maxLength: 4096, writeOnly: true }
        locale: { type: string, pattern: '^[a-z]{2}-[A-Z]{2}$', default: zh-CN }
        region: { type: [string, 'null'], pattern: '^[A-Z]{2}$' }
    TmdbConnectionTestResult:
      description: Stable, redacted outcome of a non-persistent TMDB connection test.
      type: object
      additionalProperties: false
      required: [reachable, health, failure_code, checked_at]
      properties:
        reachable: { type: boolean }
        health: { enum: [healthy, degraded, unavailable, unauthorized, rate-limited] }
        failure_code: { oneOf: [{ $ref: '#/components/schemas/IntegrationFailureCode' }, { type: 'null' }] }
        checked_at: { type: string, format: date-time }
...
    IntegrationHealth:
      description: Stable redacted health classification shared by built-in integrations.
      type: string
      enum: [unconfigured, healthy, degraded, unavailable, unauthorized, rate-limited]
    IntegrationFailureCode:
      description: Stable redacted failure classification for TMDB connection and health projections.
      type: string
      enum: [integration.not-configured, integration.unauthorized, integration.rate-limited, integration.unavailable, provider.timeout, provider.response-too-large, provider.invalid-response]
...
    AutomationFailureCode:
      description: Stable redacted source, action, and local-enhancer failure category.
      type: string
      enum: [automation.source-disabled, automation.signature-invalid, automation.replay, automation.action-invalid, automation.payload-invalid, automation.downstream-conflict, integration.not-configured, integration.unauthorized, integration.rate-limited, integration.unavailable, provider.timeout, provider.response-too-large, provider.invalid-response]
...
    IdentificationEnhancerInput:
      description: Candidate built-in Ollama configuration; only local-network endpoints are accepted by Core.
      type: object
      additionalProperties: false
      required: [enabled, base_url, model, timeout_ms]
      properties:
        enabled: { type: boolean }
        base_url: { type: string, format: uri, minLength: 1, maxLength: 2048 }
        model: { type: string, minLength: 1, maxLength: 128 }
        timeout_ms: { type: integer, minimum: 100, maximum: 30000 }
    IdentificationEnhancer:
      description: Redacted local enhancer configuration and health without prompts or responses.
      type: object
      additionalProperties: false
      required: [kind, enabled, endpoint_summary, model, timeout_ms, config_version, health, checked_at, fallback_code, projection_version, updated_at]
      properties:
        kind: { const: ollama }
        enabled: { type: boolean }
        endpoint_summary: { type: string, minLength: 1, maxLength: 255 }
        model: { type: string, minLength: 1, maxLength: 128 }
        timeout_ms: { type: integer, minimum: 100, maximum: 30000 }
        config_version: { type: integer, minimum: 1, maximum: 9007199254740991 }
        health: { $ref: '#/components/schemas/IntegrationHealth' }
        checked_at: { type: [string, 'null'], format: date-time }
        fallback_code: { oneOf: [{ $ref: '#/components/schemas/AutomationFailureCode' }, { type: 'null' }] }
        projection_version: { type: integer, minimum: 1, maximum: 9007199254740991 }
        updated_at: { type: string, format: date-time }
    IdentificationEnhancerConnectionTestResult:
      description: Non-persistent Ollama capability and health result without raw protocol bodies.
      type: object
      additionalProperties: false
      required: [reachable, health, adapter_version, model_available, fallback_code, checked_at]
      properties:
        reachable: { type: boolean }
        health: { $ref: '#/components/schemas/IntegrationHealth' }
        adapter_version: { type: string, minLength: 1, maxLength: 64 }
        model_available: { type: boolean }
        fallback_code: { oneOf: [{ $ref: '#/components/schemas/AutomationFailureCode' }, { type: 'null' }] }
        checked_at: { type: string, format: date-time }
...